# remove adler32 COPY variant (memcpy is faster)